tpm2_ptool: modularize
tpm2_ptool.py was getting quite unwieldy in size to work on, break it up
into a package and modules and use setuptools to create the CLI script.
Additionally, add a tpm2_ptool script so folks can set PYTHONPATH to the
tpm2-pkcs11/tools folder and run it. The only change people should
notice is that they need to either run:

$ python -m pip install --user -e <path to setup.py>

or

$ EXPORT PYTHONPATH=<path to tpm2-pkcs11/tools dir>:$PYTHONPATH
$ export PATH = <path to tpm2-pkcs11/tools dir>:$PATH
$ tpm2_ptool
