+++ tools/.gitignore
@@ -1,0 +1,4 @@
+*__pycache__/
+*.swp
+*.pyc
+*.egg-info/
