+++ tools/README.md
@@ -1,0 +1,21 @@
+# Bills Utilz
+
+## Install For Development
+
+```console
+git clone https://github.intel.com/johnsa1/bills-utilz
+# git clone https://github.intel.com/wcrobet/bills-utilz
+python3 -m pip install --user -e .
+```
+
+## Usage
+
+```console
+$ bill
+BILL!
+$ billbill
+BILLBILL!
+$ ufb
+It's a bird it's a plane!
+NO! It's a BILL!!
+```
