+++ tools/tpm2_pkcs11/pkcs11t.py
@@ -1,0 +1,26 @@
+# These CK* values come from the PKCS#11 spec
+CKA_KEY_TYPE = 0x100
+CKA_CLASS = 0
+
+CKO_PUBLIC_KEY = 2
+CKO_PRIVATE_KEY = 3
+CKO_SECRET_KEY  = 4
+
+CKK_RSA = 0
+CKK_EC = 0x3
+CKK_AES = 0x1f
+
+CKM_RSA_X_509=0x3
+CKM_RSA_PKCS_OAEP = 9
+CKM_AES_CBC = 0x1082
+CKM_ECDSA = 0x1041
+
+CKM_SHA256 = 0x250
+CKG_MGF1_SHA256 = 0x2
+
+CKA_LABEL = 0x3
+CKA_ID = 0x102
+CKA_MODULUS = 0x120
+CKA_PUBLIC_EXPONENT = 0x122
+CKA_VALUE_BITS = 0x160
+CKA_VALUE_LEN = 0x161
