+++ tools/tpm2_pkcs11/tpm2.py
@@ -1,0 +1,202 @@
+from __future__ import print_function
+
+import os
+import shutil
+import sys
+import tempfile
+import uuid
+import yaml
+
+from subprocess import Popen, PIPE
+
+class Tpm2(object):
+    def __init__(self, tmp, final):
+        self._tmp = tmp
+        self._final = final
+
+    def createprimary(self, ownerauth, objauth):
+        ctx = os.path.join(self._tmp, "context.out")
+        cmd = ['tpm2_createprimary',
+                    '-p', 'hex:%s' % objauth.decode(),
+                    '-o', ctx,
+                    '-g', 'sha256',
+                    '-G', 'rsa']
+
+        if ownerauth and len(ownerauth) > 0:
+            cmd.extend(['-P', ownerauth])
+
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE, env=os.environ)
+        _, stderr = p.communicate()
+        if (p.wait()):
+            raise RuntimeError("Could not execute tpm2_createprimary: %s" % stderr)
+        return ctx
+
+    @staticmethod
+    def evictcontrol(ownerauth, ctx):
+
+        cmd = ['tpm2_evictcontrol', '-c', str(ctx) ]
+
+        if ownerauth and len(ownerauth) > 0:
+            cmd.extend(['-P', ownerauth])
+
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE, env=os.environ)
+        stdout, stderr = p.communicate()
+        y=yaml.load(stdout)
+        rc = p.wait()
+        handle = y['persistentHandle'] if rc == 0 else None
+        if (p.wait()):
+            raise RuntimeError("Could not execute tpm2_evictcontrol: %s", stderr)
+        return handle
+
+    def load(self, pctx, pauth, priv, pub):
+
+        ctx = os.path.join(self._tmp, uuid.uuid4().hex + '.out')
+
+        #tpm2_load -C $file_primary_key_ctx  -u $file_load_key_pub  -r $file_load_key_priv -n $file_load_key_name -o $file_load_key_ctx
+        cmd = ['tpm2_load', '-C', str(pctx), '-P', 'hex:' + pauth.decode(),'-u', pub, '-r', priv, '-n', '/dev/null', '-o', ctx]
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE, env=os.environ)
+        _, stderr = p.communicate()
+        rc = p.wait()
+        if rc:
+            raise RuntimeError("Could not execute tpm2_load: %s", stderr)
+        return ctx
+
+    def unseal(self, ctx, auth):
+
+        # tpm2_unseal -Q -c $file_unseal_key_ctx
+        cmd = ['tpm2_unseal', '-c', ctx, '-p', 'hex:'+auth.decode()]
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE, env=os.environ)
+        stdout, stderr = p.communicate()
+        rc = p.wait()
+        if rc:
+            raise RuntimeError("Could not execute tpm2_unseal: %s", stderr)
+        return stdout
+
+    def _encryptdecrypt(self, ctx, auth, data, decrypt=False):
+
+        cmd = ['tpm2_encryptdecrypt', '-c', ctx, '-p', 'hex:' + auth.decode()]
+
+        if decrypt:
+            cmd.extend(['-D'])
+
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE, stdin=PIPE, env=os.environ)
+        stdout, stderr = p.communicate(input=data)
+        rc = p.wait()
+        if rc:
+            raise RuntimeError("Could not execute tpm2_encryptdecrypt: %s", stderr)
+        return stdout
+
+    def encrypt(self, ctx, auth, data):
+        return self._encryptdecrypt(ctx, auth, data)
+
+    def decrypt(self, ctx, auth, data):
+        return self._encryptdecrypt(ctx, auth, data, decrypt=True)
+
+    def _move(self, path):
+        b = os.path.basename(path)
+        n = os.path.join(self._final, b)
+        shutil.move(path, n)
+        return n
+
+    def create(self, phandle, pauth, objauth, objattrs=None, seal=None, alg=None):
+        # tpm2_create -Q -C context.out -g $gAlg -G $GAlg -u key.pub -r key.priv
+        _, priv = tempfile.mkstemp(prefix='', suffix='.priv', dir=self._tmp)
+        _, pub = tempfile.mkstemp(prefix='', suffix='.pub', dir=self._tmp)
+
+        cmd = ['tpm2_create', '-C', str(phandle), '-u', pub, '-r', priv]
+
+        if pauth and len(pauth) > 0:
+            cmd.extend(['-P', 'hex:%s' % pauth.decode()])
+
+        if objauth and len(objauth) > 0:
+            cmd.extend(['-p', 'hex:%s' % objauth.decode()])
+
+        if objattrs != None:
+            cmd.extend(['-A', objattrs])
+
+        if seal != None:
+            cmd.extend(['-I', '-'])
+
+        if alg != None:
+            cmd.extend(['-G', alg])
+
+        p = Popen(cmd,
+                   stdout=PIPE, stderr=PIPE, stdin=PIPE, env=os.environ)
+        stdout, stderr = p.communicate(input=seal)
+        rc = p.wait()
+        if (rc != 0):
+            os.remove(pub)
+            os.remove(priv)
+            raise RuntimeError("Could not execute tpm2_create: %s" % str(stderr))
+
+        pub = self._move(pub)
+        priv = self._move(priv)
+        return priv, pub, stdout
+
+    def getcap(self, cap):
+
+        # tpm2_getcap -Q -l $cap
+        cmd = ['tpm2_getcap', '-c', cap]
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE, env=os.environ)
+        stdout, stderr = p.communicate()
+        rc = p.wait()
+        if rc:
+            raise RuntimeError("Could not execute tpm2_getcap: %s", stderr)
+        return stdout
+
+    def importkey(self, phandle, pauth, objauth, privkey, objattrs=None, seal=None, alg=None):
+
+        _, priv = tempfile.mkstemp(prefix='', suffix='.priv', dir=self._tmp)
+        _, pub = tempfile.mkstemp(prefix='', suffix='.pub', dir=self._tmp)
+
+        if privkey and len(privkey) > 0:
+            exists = os.path.isfile(privkey)
+            if not exists:
+                raise RuntimeError("File '%s' path is invalid or is missing", privkey)
+        else:
+            sys.exit("Invalid file path")
+
+        parent_path = "file:"+ str(phandle)
+        cmd = ['tpm2_import', '-V', '-C', parent_path, '-k', privkey, '-u', pub, '-r', priv]
+
+        if pauth and len(pauth) > 0:
+            cmd.extend(['-P', 'hex:%s' % pauth.decode()])
+
+        if objauth and len(objauth) > 0:
+            cmd.extend(['-p', 'hex:%s' % objauth.decode()])
+
+        if objattrs != None:
+            cmd.extend(['-A', objattrs])
+
+        if seal != None:
+            cmd.extend(['-I', '-'])
+
+        if alg != None:
+            cmd.extend(['-G', alg])
+
+        p = Popen(cmd,
+                   stdout=PIPE, stderr=PIPE, stdin=PIPE, env=os.environ)
+        stdout, stderr = p.communicate(input=seal)
+        rc = p.wait()
+        if (rc != 0):
+            os.remove(pub)
+            os.remove(priv)
+            raise RuntimeError("Could not execute tpm2_import: %s" % str(stderr))
+
+        pub = self._move(pub)
+        priv = self._move(priv)
+        return priv, pub, stdout
+
+    def changeauth(self, pctx, objctx, oldobjauth, newobjauth):
+
+        newpriv = os.path.join(self._tmp, uuid.uuid4().hex + '.priv')
+
+        #tpm2_load -C $file_primary_key_ctx  -u $file_load_key_pub  -r $file_load_key_priv -n $file_load_key_name -o $file_load_key_ctx
+        cmd = ['tpm2_changeauth', '-a', str(pctx), '-c', str(objctx), '-P', 'hex:' + oldobjauth.decode(), '-p', 'hex:' + newobjauth.decode(), '-r', newpriv]
+        p = Popen(cmd, stdout=PIPE, stderr=PIPE, env=os.environ)
+        _, stderr = p.communicate()
+        rc = p.wait()
+        if rc:
+            raise RuntimeError("Could not execute tpm2_load: %s", stderr)
+        newpriv = self._move(newpriv)
+        return newpriv
