+++ tools/tpm2_ptool
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+python -m tpm2_pkcs11.tpm2_ptool $@
